[Refactor] manage image workflow scripts [Feat.] add series workflow
