+++ fetch_drive.py
@@ -1,6 +1,42 @@
 from pydrive.auth import GoogleAuth
 from pydrive.drive import GoogleDrive
 import os
+import time
+from datetime import datetime
+def check_task_status(task, gap = 20):
+    """
+    监控任务状态直到完成或失败
+    
+    Args:
+        task: ee.batch.Task 对象
+        timeout_minutes: 超时时间（分钟）
+    
+    Returns:
+        bool: 任务是否成功完成
+    """
+    
+    while True:
+        # 获取任务状态
+        status = task.status()
+        state = status['state']
+        
+        # 打印当前状态
+        print(f"[{datetime.now().strftime('%Y-%m-%d %H:%M:%S')}] 任务状态: {state}")
+        
+        # 检查是否完成
+        if state == 'COMPLETED':
+            print("✓ 任务成功完成！")
+            return True
+            
+        # 检查是否失败
+        elif state in ['FAILED', 'CANCELLED']:
+            error_message = status.get('error_message', '未知错误')
+            print(f"× 任务失败: {error_message}")
+            return False
+            
+        # 等待一段时间再检查
+        time.sleep(gap)  # 10秒检查一次
+
 
 def create_folder(drive,parent_folder_id,folder_name):
     folder_metadata = {
@@ -27,11 +63,10 @@ def get_folder_id_by_name(drive, folder_name, parent_id='root'):
         return file_list[0]['id']
     return None
 
-def download_and_clean(drive,folder_name, save_path):
+def download_and_clean(drive,folder_id, save_path):
     # 创建保存目录
     os.makedirs(save_path, exist_ok=True)
     
-    folder_id = get_folder_id_by_name(drive,folder_name)
     # 查询文件夹内容（排除子文件夹）
     file_list = drive.ListFile({
         'q': f"'{folder_id}' in parents and mimeType != 'application/vnd.google-apps.folder'",
